Upon DNS failure, fallback to removing SSID from encoded string
When testing with an open SSID of "ASeriesOfTubes" the DNS request always failed, but worked with an SSID of "ASeries", likely due to the length of the subdomain hitting the ceiling of what the WiFi.hostByName() is able to handle.  To solve this, a dirty hack of 2 failures tries a 3rd time without appending the SSID to the GPS data.  Lab testing showed this as a successful fallback approach.